Move kunenaversion into libraries\version and kunenamodelinstall in inst
